Add files via upload
This script analyses the impact of air filtration on pathogen detection in an interventional substudy in a childcare facility. 
Regarding the generated Figure (Fig 2 in the paper), small post-hoc changes have been made in the final version in the paper. These include the addition of lines highlighting significant differences and colour changes.